move rolling_file_appender into its own subpackage
Make messageFmt and args in Log struct public
